requirements.txt
backports.tarfile==1.2.0
confluent-kafka==2.10.0
djangorestframework==3.16.0
grpcio-tools==1.73.0
gunicorn==23.0.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
pip-chill==1.0.3
platformdirs==4.2.2
tomli==2.0.1
